progress 3

+++ app/src/main/java/com/ultimo/vehicleapp/model/Material.kt
@@ -7,5 +7,5 @@ data class Material_List(
     val bahan_id: Int,
     val nama_bahan: String,
     val deskripsi: String?,
-    val harga_per_unit: Double,
+    val harga_per_unit: Int,
 )
